配置测试环境
