+++ src/main/java/com/hitbd/proj/Main.java
@@ -15,7 +15,7 @@ public class Main {
             System.out.println("Usage: trafficBD Action [Parameter]");
             System.out.println("Actions:");
             System.out.println("ImportAlarm FileName/Folder HBaseConfFile   #import alarm file to hbase");
-            System.out.println("ImportDevice FileName/Folder import   #alarm file to ignite");
+            System.out.println("ImportDevice FileName/Folder  #import device file to ignite");
             System.out.println("CreateIgniteTable [hostname]  #create user_b, user_c and device table");
             return;
         }
