+++ src/test/java/com/hitbd/proj/HbaseTest.java
@@ -2,33 +2,22 @@ package com.hitbd.proj;
 
 import com.hitbd.proj.logic.AlarmScanner;
 import com.hitbd.proj.model.Pair;
-import org.junit.Assert;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.junit.Ignore;
 import org.junit.Test;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Date;
 
-public class IgniteTest {
+public class HbaseTest {
     @Ignore
     @Test
-    public void testCache() {
-        /*
-         * 测试AlarmC和ViewedC是否可以正确执行
-         * 已经完成的测试，出于性能原因，在打包时进行忽略
-         */
-        IgniteSearch search = IgniteSearch.getInstance();
-        search.setAlarmCount(123L, 2);
-        Assert.assertEquals(search.getAlarmCount(123L), 2);
-        search.setAlarmCount(456L, 456);
-        Assert.assertEquals(search.getAlarmCount(456L), 456);
-        search.setViewedCount(123L, 1);
-        Assert.assertEquals(search.getViewedCount(123L), 1);
-    }
-
-    @Ignore
-    @Test
-    public void testScanner() {
+    public void testQuery() throws IOException{
+        Configuration configuration = HBaseConfiguration.create();
+        configuration.addResource("/usr/hbase-1.3.2.1/conf/hbase-site.xml");
+        Settings.HBASE_CONFIG = configuration;
         ArrayList<Integer> userBIds = new ArrayList<>();
         userBIds.add(2469);
         QueryFilter queryFilter = new QueryFilter();
@@ -36,5 +25,10 @@ public class IgniteTest {
         Date date = new Date();
         AlarmScanner scanner = HbaseSearch.getInstance().queryAlarmByUser(2469, userBIds, true, HbaseSearch.NO_SORT, queryFilter);
         System.out.println(" Use Time:" + (new Date().getTime() - date.getTime()) + "ms; create query " + scanner.queries.size());
+        date = new Date();
+        while (!scanner.isFinished()) {
+            scanner.next();
+        }
+        System.out.println("Hbase Use Time:" + (new Date().getTime() - date.getTime()) + "ms; scan alarm: " + scanner.getTotalAlarm());
     }
 }
